rename fconf
