+++ tests/fconf_data/1.in/1.yml
@@ -1,0 +1,6 @@
+---
+templates:
+  blue(one): ${one} for you
+
+one:
+  two: three('hello')
