--- tests/pkyaml1_test.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-u"""test parse_files conformance
-
-:copyright: Copyright (c) 2022 RadiaSoft LLC.  All Rights Reserved.
-:license: http://www.apache.org/licenses/LICENSE-2.0.html
-"""
-import pytest
