--- tests/pkyaml_data/1.in/1.yml
@@ -1,3 +1,0 @@
----
-one:
-  two: three('hello')
